pod-control/requirements.txt
python-socketio==5.7.2
uvicorn[standard]==0.20.0
